[style] : Lotto -> Lottos, LottoNumber -> Lotts class name 변경
