+++ src/main/java/edu/nextstep/step4/domain/Lotto.java
@@ -1,8 +1,10 @@
 package edu.nextstep.step4.domain;
 
-import edu.nextstep.step4.enums.Rank;
-
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 /**
  * author       : gwonbyeong-yun <sksggg123>
@@ -12,44 +14,58 @@ import java.util.*;
  * | blog         : sksggg123.github.io     |
  * ------------------------------------------
  * project      : java-lotto
- * create date  : 2019-06-25 00:02
+ * create date  : 2019-06-24 23:28
  */
 public class Lotto {
 
-    private List<LottoNumber> lotto;
+    private static final int LIST_VALID_SIZE = 6;
+    private static final String COMMA = ",";
+    private static final String SPACE = " ";
+    private List<Number> lottoNumbers;
 
-    public Lotto(List<LottoNumber> lotto) {
-        this.lotto = new ArrayList<>(lotto);
+    public Lotto(List<Number> lottoNumbers) {
+        validSize(lottoNumbers);
+        this.lottoNumbers = new ArrayList<>(lottoNumbers);
     }
 
-    public Map<LottoNumber, Rank> getLotteryLottoNumberResultCount(LottoNumber lottery) {
-        Map<LottoNumber, Rank> lottoResultinfo = new HashMap<>();
-        lotto.stream()
-                .forEach(lottoNumber ->
-                        lottoResultinfo.put(lottoNumber, Rank.matchCheck(lottoNumber.compareMatchNumberCount(lottery)))
-                );
-        return lottoResultinfo;
+    public static Lotto createLotteryNumber(String inputLotteryNumber) {
+        return new Lotto(splitLotteryNumber(inputLotteryNumber));
     }
 
-    public Map<LottoNumber, Rank> addBonusNumberMatchLotto(Map<LottoNumber, Rank> lotteryInfo, Number bonusNumber) {
-        Map<LottoNumber, Rank> cloneInfoMap = new HashMap<>(lotteryInfo);
-
-        cloneInfoMap.keySet().stream()
-                .filter(lotto -> cloneInfoMap.get(lotto).getMatch() == Rank.FIVE.getMatch())
-                .filter(lotto -> lotto.contains(bonusNumber.getNumber()))
-                .forEach(lotto ->
-                        cloneInfoMap.put(lotto, Rank.BONUS)
-                );
-        return cloneInfoMap;
+    public boolean contains(int number) {
+        return lottoNumbers.stream()
+                .anyMatch(sourceNumber -> sourceNumber.comapreTo(number));
     }
 
-    public int fetchIncome(Map<LottoNumber, Rank> lotteryInfo) {
-        return lotteryInfo.keySet().stream()
-                .mapToInt(lotto -> lotteryInfo.get(lotto).getLotteryMoney())
-                .sum();
+    public int compareMatchNumberCount(Lotto lotto) {
+        return lotto.stream()
+                .filter(compareSourceNumber -> this.contains(compareSourceNumber.getNumber()))
+                .collect(Collectors.toList())
+                .size();
     }
 
-    public List<LottoNumber> getExtractOfLotto() {
-        return Collections.unmodifiableList(lotto);
+    public Stream<Number> stream() {
+        return this.lottoNumbers.stream();
     }
+
+    private void validSize(List<Number> lotteryNumber) {
+        int numbers = lotteryNumber.stream()
+                .map(Number::getNumber)
+                .distinct()
+                .collect(Collectors.toList())
+                .size();
+        if (numbers != LIST_VALID_SIZE) {
+            throw new IllegalArgumentException("전달된 숫자가 6개가 아닙니다.");
+        }
+    }
+
+    private static List<Number> splitLotteryNumber(String lottery) {
+
+        return Arrays.stream(lottery.replaceAll(SPACE, "").split(COMMA))
+                .filter(splitString -> !"".equals(splitString))
+                .map(Integer::parseInt)
+                .map(number -> new Number(number))
+                .collect(Collectors.toList());
+    }
+
 }
