+++ src/main/java/edu/nextstep/step4/view/ResultView.java
@@ -1,8 +1,8 @@
 package edu.nextstep.step4.view;
 
 import edu.nextstep.step4.enums.Rank;
+import edu.nextstep.step4.domain.Lottos;
 import edu.nextstep.step4.domain.Lotto;
-import edu.nextstep.step4.domain.LottoNumber;
 import edu.nextstep.step4.domain.Number;
 
 import java.util.Arrays;
@@ -29,8 +29,8 @@ public class ResultView {
     private static final String BONUS_MESSAGE_FORMAT = "%d개 일치, 보너스 볼 일치 (%d원)- %d개";
     private static final String INCOME_MESSAGE_FORMAT = "총 수익률은 %.2f입니다. (기준이 1이기 때문에 결과적으로 손해라는 의미임)";
 
-    public static void printLottoInfo(Lotto lotto) {
-        List<String> extractInfos = lotto.getExtractOfLotto().stream()
+    public static void printLottoInfo(Lottos lottos) {
+        List<String> extractInfos = lottos.getExtractOfLotto().stream()
                 .map(extractionNumber -> combineExtractNumber(extractionNumber))
                 .collect(Collectors.toList());
 
@@ -42,7 +42,7 @@ public class ResultView {
         System.out.println(String.format(INCOME_MESSAGE_FORMAT, income));
     }
 
-    public static void printLotteryCount(Map<LottoNumber, Rank> lotteryInfo) {
+    public static void printLotteryCount(Map<Lotto, Rank> lotteryInfo) {
         System.out.println("\n당첨 통계\n-------");
         Arrays.stream(Rank.values())
                 .filter(ranks -> ranks != Rank.ZERO)
@@ -58,10 +58,10 @@ public class ResultView {
         return String.format(MESSAGE_FORMAT, count, money, matchCount);
     }
 
-    private static String combineExtractNumber(LottoNumber lottoNumber) {
+    private static String combineExtractNumber(Lotto lotto) {
         StringBuilder sb = new StringBuilder();
         sb.append(PREFIX);
-        List<Integer> numbers = lottoNumber.stream()
+        List<Integer> numbers = lotto.stream()
                 .map(Number::getNumber)
                 .collect(Collectors.toList());
