+++ src/test/java/edu/nextstep/step4/domain/LottoTest.java
@@ -1,16 +1,20 @@
 package edu.nextstep.step4.domain;
 
-import edu.nextstep.step3.Rank;
-import edu.nextstep.step3.domain.Lotto;
 import edu.nextstep.step3.domain.LottoNumber;
 import edu.nextstep.step3.domain.Number;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.ValueSource;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
+import java.util.stream.Collectors;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 
 /**
  * author       : gwonbyeong-yun <sksggg123>
@@ -20,70 +24,76 @@ import static org.assertj.core.api.Assertions.assertThat;
  * | blog         : sksggg123.github.io     |
  * ------------------------------------------
  * project      : java-lotto
- * create date  : 2019-06-24 23:53
+ * create date  : 2019-06-24 23:24
  */
 public class LottoTest {
 
     private List<Number> numbers;
-    private List<Number> lotteryNumbers;
 
     @BeforeEach
     void setUp() {
-        numbers = new ArrayList<>();
-        numbers.add(new Number(1));
-        numbers.add(new Number(2));
-        numbers.add(new Number(3));
-        numbers.add(new Number(4));
+        Number number1 = new Number(1);
+        Number number2 = new Number(2);
+        Number number3 = new Number(3);
+        Number number4 = new Number(4);
+        Number number5 = new Number(5);
+        numbers = new ArrayList<>(Arrays.asList(number1, number2, number3, number4, number5));
+
+    }
+
+    @DisplayName("리스트 size가 6으로 전달받아 생성하기 - 예외상황")
+    @Test
+    void validSize() {
+        assertThatIllegalArgumentException().isThrownBy(() -> {
+            LottoNumber exNumber = new LottoNumber(numbers);
+        }).withMessageContaining("전달된 숫자가 6개가 아닙니다.");
+    }
+
+    @DisplayName("리스트에 중복된 숫자 제거하기")
+    @Test
+    void validDuplicate() {
         numbers.add(new Number(5));
+        assertThatIllegalArgumentException().isThrownBy(() -> {
+            LottoNumber exNumber = new LottoNumber(numbers);
+        }).withMessageContaining("전달된 숫자가 6개가 아닙니다.");
+    }
+
+    @DisplayName("전달된 인자값(int type)을 보유하고 있는지 확인한다.")
+    @ParameterizedTest
+    @ValueSource(ints = {1, 2, 3, 4, 5, 6})
+    void contains(int number) {
         numbers.add(new Number(6));
-
-        lotteryNumbers = new ArrayList<>();
-        lotteryNumbers.add(new Number(1));
-        lotteryNumbers.add(new Number(2));
-        lotteryNumbers.add(new Number(3));
-        lotteryNumbers.add(new Number(4));
-        lotteryNumbers.add(new Number(5));
-        lotteryNumbers.add(new Number(7));
+        LottoNumber lottoNumber = new LottoNumber(numbers);
+        assertThat(lottoNumber.contains(number)).isTrue();
     }
 
-    @DisplayName("발행된 로또번호들과 당첨로또를 비교하여 Map에 LottoNumber와 Rank를 key, value로 반환")
+    @DisplayName("ExtractionNumber 객체데이터의 당첨된 카운트 반환는다.")
     @Test
-    void getLotteryLottoNumberResultCount() {
-        LottoNumber lottoNumber = new LottoNumber(new ArrayList<>(numbers));
-        LottoNumber lotteryNumber = new LottoNumber(new ArrayList<>(numbers));
-        Lotto lotto = new Lotto(Arrays.asList(lottoNumber));
+    void contains() {
+        numbers.add(new Number(6));
+        LottoNumber lottoNumber = new LottoNumber(numbers);
 
-        assertThat(lotto.getLotteryLottoNumberResultCount(lotteryNumber).get(lottoNumber)).isEqualTo(Rank.SIX);
+        List<Number> compareNumber = new ArrayList<>();
+
+        compareNumber.add(new Number(2));
+        compareNumber.add(new Number(4));
+        compareNumber.add(new Number(6));
+        compareNumber.add(new Number(8));
+        compareNumber.add(new Number(10));
+        compareNumber.add(new Number(12));
+        LottoNumber exNumber = new LottoNumber(compareNumber);
+
+        assertThat(lottoNumber.compareMatchNumberCount(exNumber)).isEqualTo(3);
     }
 
-    @DisplayName("5개 당첨된 로또들 중 보너스 번호와 비교하여 Map에 LottoNumber와 Rank를 key, value로 반환")
-    @Test
-    void addBonusNumberMatchLotto() {
-        Number bonus = new Number(6);
-        LottoNumber lottoNumber = new LottoNumber(new ArrayList<>(numbers));
-        LottoNumber lotteryNumber = new LottoNumber(new ArrayList<>(lotteryNumbers));
-        Lotto lotto = new Lotto(Arrays.asList(lottoNumber));
-
-        Map<LottoNumber, Rank> lotteryInfo = new HashMap<>();
-        lotteryInfo.put(lottoNumber, Rank.matchCheck(lottoNumber.compareMatchNumberCount(lotteryNumber)));
-
-        assertThat(lotto.addBonusNumberMatchLotto(lotteryInfo, bonus).get(lottoNumber)).isEqualTo(Rank.BONUS);
-    }
-
-    @DisplayName("당첨 금액 반환")
-    @Test
-    void fetchIncome() {
-        Number bonus = new Number(6);
-        LottoNumber lottoNumber = new LottoNumber(new ArrayList<>(numbers));
-        LottoNumber lotteryNumber = new LottoNumber(new ArrayList<>(lotteryNumbers));
-        Lotto lotto = new Lotto(Arrays.asList(lottoNumber));
-
-        Map<LottoNumber, Rank> lotteryInfo = new HashMap<>();
-        lotteryInfo.put(lottoNumber, Rank.matchCheck(lottoNumber.compareMatchNumberCount(lotteryNumber)));
-
-        Map<LottoNumber, Rank> incomeInfo = lotto.addBonusNumberMatchLotto(lotteryInfo, bonus);
-        lotteryInfo.put(lottoNumber, Rank.matchCheck(lottoNumber.compareMatchNumberCount(lotteryNumber)));
-
-        assertThat(lotto.fetchIncome(incomeInfo)).isEqualTo(30_000_000);
+    @DisplayName("당첨번호 생성하는 기능")
+    @ParameterizedTest
+    @ValueSource(ints = {1, 2, 3, 4, 5, 6})
+    void createLotteryNumber(int number) {
+        LottoNumber lottoNumber = LottoNumber.createLotteryNumber("1,2,3,4,5,6");
+        assertThat(lottoNumber.stream()
+                .map(Number::getNumber)
+                .collect(Collectors.toList()).contains(number)
+        );
     }
 }
